drf/gb/requirements.txt
Django==4.0
djangorestframework==3.12.4
psycopg2-binary==2.9.2
markdown==3.3.4
django-filter==21.1
django-cors-headers==3.10.1
pytz
mixer==7.2.0
django-extensions==3.1.5